requirements.txt
Django==5.0.2
djangorestframework==3.15.2
phonenumbers==8.13.43
PyJWT==2.3.0
psycopg==3.2.1
psycopg2-binary==2.9.9
django-cors-headers==4.4.0
persian==0.5.0
requests==2.32.3
cryptography==43.0.0
pytz==2024.1
jdatetime==5.0.0